CSQ mapping working. Fixed an issue with DatHandler. Fixed digamma funcs
