+++ src/DatObject/Make_Dat.py
@@ -3,9 +3,9 @@ make DatHDFs"""
 from __future__ import annotations
 import os
 import logging
+from singleton_decorator import singleton
 
 from src.DatObject.DatHDF import DatHDF, get_dat_id, DatHDFBuilder
-from singleton_decorator import singleton
 import src.HDF_Util as HDU
 from typing import TYPE_CHECKING, Union, Iterable, Tuple, List
 import threading
@@ -26,7 +26,7 @@ logger = logging.getLogger(__name__)
 sync_lock = threading.Lock()
 
 
-@singleton  # Necessary when only calling on class variables anyway?
+@singleton
 class DatHandler(object):
     """
     Holds onto references to open dats (so that I don't try open the same datHDF more than once). Will return
@@ -35,36 +35,34 @@ class DatHandler(object):
     """
     open_dats = {}
 
-    @classmethod
-    def get_dat(cls, datnum: int, datname='base', overwrite=False,  init_level='min', exp2hdf: type(Exp2HDF) = None) \
+    def get_dat(self, datnum: int, datname='base', overwrite=False,  init_level='min', exp2hdf: type(Exp2HDF) = None) \
             -> DatHDF:
         exp2hdf = exp2hdf(datnum=datnum, datname=datname) if exp2hdf else \
             default_Exp2HDF(datnum=datnum, datname=datname)
         full_id = f'{exp2hdf.ExpConfig.dir_name}:{get_dat_id(datnum, datname)}'  # For temp local storage
         path = exp2hdf.get_datHDF_path()
-        cls._ensure_dir(path)
+        self._ensure_dir(path)
         if overwrite:
-            cls._delete_hdf(path)
-            if full_id in cls.open_dats:
-                del cls.open_dats[full_id]
+            self._delete_hdf(path)
+            if full_id in self.open_dats:
+                del self.open_dats[full_id]
 
-        if full_id not in cls.open_dats:  # Need to open or create DatHDF
+        if full_id not in self.open_dats:  # Need to open or create DatHDF
             if os.path.isfile(path):
-                cls.open_dats[full_id] = cls._open_hdf(path)
+                self.open_dats[full_id] = self._open_hdf(path)
             else:
-                cls._check_exp_data_exists(exp2hdf)
+                self._check_exp_data_exists(exp2hdf)
                 builder = DatHDFBuilder(exp2hdf, init_level)
-                cls.open_dats[full_id] = builder.build_dat()
-        return cls.open_dats[full_id]
+                self.open_dats[full_id] = builder.build_dat()
+        return self.open_dats[full_id]
 
-    @classmethod
-    def get_dats(cls, datnums: Union[Iterable[int], Tuple[int, int]], datname='base', overwrite=False, init_level='min',
+    def get_dats(self, datnums: Union[Iterable[int], Tuple[int, int]], datname='base', overwrite=False, init_level='min',
                  exp2hdf=None) -> List[DatHDF]:
         """Convenience for loading multiple dats at once, just calls get_dat multiple times"""
         # TODO: Make this multiprocess/threaded especially if overwriting or if dat does not already exist!
         if type(datnums) == tuple and len(datnums) == 2:
             datnums = range(*datnums)
-        return [cls.get_dat(num, datname=datname, overwrite=overwrite, init_level=init_level, exp2hdf=exp2hdf)
+        return [self.get_dat(num, datname=datname, overwrite=overwrite, init_level=init_level, exp2hdf=exp2hdf)
                 for num in datnums]
 
     @staticmethod
@@ -101,13 +99,11 @@ class DatHandler(object):
                     return True
                 raise FileNotFoundError(f'No experiment data found for dat{exp2hdf.datnum} at {os.path.abspath(exp_path)}')
 
-    @classmethod
-    def list_open_dats(cls):
-        return cls.open_dats
+    def list_open_dats(self):
+        return self.open_dats
 
-    @classmethod
-    def clear_dats(cls):
-        cls.open_dats = {}
+    def clear_dats(self):
+        self.open_dats = {}
 
 
 get_dat = DatHandler().get_dat
